+++ sphinxcontrib/redoc.py
@@ -29,13 +29,6 @@ def render(app):
         # Setup options if they are not passed since 'redoc.j2' template
         # relies on them.
         ctx.setdefault('opts', {})
-        ctx['opts'].setdefault('lazy-rendering', False)
-        ctx['opts'].setdefault('suppress-warnings', False)
-        ctx['opts'].setdefault('hide-hostname', False)
-        ctx['opts'].setdefault('required-props-first', False)
-        ctx['opts'].setdefault('no-auto-auth', False)
-        ctx['opts'].setdefault('path-in-middle-panel', False)
-        ctx['opts'].setdefault('expand-responses', [])
 
         # The 'spec' may contain either HTTP(s) link or filesystem path. In
         # case of later we need to copy the spec into output directory, as